requirements.txt
Flask==2.3.2
Pillow==9.4.0
paddlepaddle==2.5.0
paddleocr==2.6.1.3
opencv-python==4.7.0.72
